+++ client/src/components/forms/DepartmentForm.tsx
@@ -2,23 +2,19 @@
 
 import { BreadcrumbLink } from "@/src/components/ui/breadcrumbs/BreadcrumbLink";
 import Button from "@/src/components/ui/button";
-import Dropdown from "@/src/components/ui/dropdown";
-import Field from "@/src/components/ui/field";
-import { useEffect } from "react";
-import { Controller, useForm } from "react-hook-form";
-
-type Mode = "create" | "edit" | "view";
-
-export type DepartmentFormData = {
-    name: string;
-    employee_id: string;
-};
+import { defaultDepartmentValues } from "@/src/lib/forms/department/department-form-default-values";
+import { Mode } from "@/src/types/form";
+import { FormData } from "@/src/types/forms/department-form.types";
+import { useEffect, useMemo } from "react";
+import { FormProvider, useForm } from "react-hook-form";
+import FormBuilder from "../layout/FormBuilder";
+import { DepartmentFormFields } from "@/src/lib/forms/department/department-form-field";
 
 type DepartmentFormProps = {
     mode: Mode;
     title: string;
-    initialValues?: Partial<DepartmentFormData>;
-    onSubmit?: (data: DepartmentFormData) => void;
+    initialValues?: Partial<FormData>;
+    onSubmit?: (data: FormData) => void;
     cancelPath?: string;
     submitLabel?: string;
 };
@@ -32,29 +28,21 @@ const DepartmentForm = ({
     submitLabel,
 }: DepartmentFormProps) => {
     const isReadOnly = mode === "view";
+    const defaultValues: FormData = useMemo(
+        () => defaultDepartmentValues(initialValues),
+        [initialValues]
+    );
 
-    const {
-        register,
-        control,
-        handleSubmit,
-        reset,
-        formState: { errors },
-    } = useForm<DepartmentFormData>({
-        defaultValues: {
-            name: initialValues?.name ?? "",
-            employee_id: initialValues?.employee_id ?? "",
-        },
+    const methods = useForm<FormData>({
+        defaultValues,
         mode: "onBlur",
     });
 
     useEffect(() => {
-        reset({
-            name: initialValues?.name ?? "",
-            employee_id: initialValues?.employee_id ?? "",
-        });
-    }, [initialValues, reset]);
+        methods.reset(defaultValues);
+    }, [defaultValues, methods]);
 
-    const handleFormSubmit = handleSubmit((data) => {
+    const handleFormSubmit = methods.handleSubmit((data) => {
         if (isReadOnly) return;
         onSubmit?.(data);
     });
@@ -68,66 +56,9 @@ const DepartmentForm = ({
                 onSubmit={handleFormSubmit}
             >
                 <div className="w-full bg-white p-4 rounded-lg shadow-md flex flex-col overflow-y-auto max-h-screen min-h-125">
-                    <div className="flex flex-col md:flex-row gap-2">
-                        <div className="w-full md:w-1/2 flex flex-col">
-                            <p className="text-sm">Name</p>
-                            <div className="flex flex-col">
-                                <Field
-                                    variant="outlined"
-                                    // color="#565555"
-                                    isRounded
-                                    disabled={isReadOnly}
-                                    {...register("name", {
-                                        required: isReadOnly
-                                            ? false
-                                            : "Name is required",
-                                    })}
-                                />
-                                {errors.name?.message && (
-                                    <p className="text-sm text-right pr-3 text-red-400">
-                                        {errors.name.message}
-                                    </p>
-                                )}
-                            </div>
-                        </div>
-                        <div className="w-full md:w-1/2 flex flex-col">
-                            <p className="text-sm">Email</p>
-                            <div className="flex flex-col">
-                                <Controller
-                                    name="employee_id"
-                                    control={control}
-                                    rules={{
-                                        required: isReadOnly
-                                            ? false
-                                            : "Email is required",
-                                    }}
-                                    render={({ field }) => (
-                                        <Dropdown
-                                            value={field.value}
-                                            onChange={field.onChange}
-                                            isRounded
-                                            disabled={isReadOnly}
-                                            options={[
-                                                {
-                                                    label: "pj.judan@flowmetricaccounting.group",
-                                                    value: 1,
-                                                },
-                                                {
-                                                    label: "judyann.vibal@flowmetricaccounting.group",
-                                                    value: 2,
-                                                },
-                                            ]}
-                                        />
-                                    )}
-                                />
-                                {errors.employee_id?.message && (
-                                    <p className="text-sm text-right pr-3 text-red-400">
-                                        {errors.employee_id.message}
-                                    </p>
-                                )}
-                            </div>
-                        </div>
-                    </div>
+                    <FormProvider {...methods}>
+                        <FormBuilder fields={DepartmentFormFields} isReadOnly={isReadOnly} />
+                    </FormProvider>
                 </div>
                 <div className="flex flex-row items-center justify-end gap-x-3">
                     <BreadcrumbLink