+++ client/src/types/datepicker.d.ts
@@ -1,0 +1,10 @@
+import { InputHTMLAttributes } from "react";
+
+export type DatePickerProps = InputHTMLAttributes<HTMLInputElement> & {
+    variant?: "outlined" | "filled";
+    isRounded?: boolean;
+    bgColor?: string;
+    color?: string;
+    minDate?: string;
+    maxDate?: string;
+};