+++ client/src/types/forms/department-form.types.ts
@@ -1,0 +1,4 @@
+export type FormData = {
+    name: string;
+    employee_id: string;
+};