+++ client/src/types/forms/service-form.types.ts
@@ -1,0 +1,4 @@
+export type FormData = {
+    name: string;
+    department_id: string;
+}